A missed conversion from int to ssize_t. (np)
git-svn-id: svn://mielke.cc/main/brltty@5376 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/usb_linux.c
@@ -627,7 +627,7 @@ usbReadEndpoint (
   size_t length,
   int timeout
 ) {
-  int count = -1;
+  ssize_t count = -1;
   UsbEndpoint *endpoint;
 
   if ((endpoint = usbGetInputEndpoint(device, endpointNumber))) {
